made the program more interactive

+++ antlr_setup.py
@@ -6,23 +6,33 @@ from platform import system as platform_os
 current_dir = getcwd()
 
 def download_antlr_jar():
+    if path.exists('./antlr4.jar'):
+        return
     url = 'https://www.antlr.org/download/antlr-4.7.2-complete.jar'
     req = urlopen(url)
     f = open('antlr4.jar', 'wb')
-    f.write(req.read())
+    total = 0
+    while True:
+        chunk = req.read(1024)
+        total += len(chunk)
+        print('\rdownloaded %d bytes' % total, end='')
+        if not chunk:
+            break
+        f.write(chunk)
     f.close()
     req.close()
 
-def run_antlr():
-    cp_string = '.;%s/antlr4.jar;%%CLASSPATH%%' if platform_os() == 'Windows' else '.:%s/antlr4.jar:%CLASSPATH%'
-    environ['CLASSPATH'] = cp_string % current_dir
-    print( environ['CLASSPATH'] )
+def run_antlr(grammar_file):
     chdir('./parser')
-    system('java -jar %s/antlr4.jar ./Definition.g4 -no-listener -visitor' % current_dir)
+    system('java -jar %s/antlr4.jar ./%s.g4 -no-listener -visitor' % (current_dir, grammar_file))
     chdir('..')
 
 def compile_java():
+    cp_string = '.;%s/antlr4.jar;%%CLASSPATH%%' if platform_os() == 'Windows' else '.:%s/antlr4.jar:%CLASSPATH%'
+    environ['CLASSPATH'] = cp_string % current_dir
     system('javac -d ./build ./parser/*.java ./*.java')
+
+def run_java():
     chdir('./build')
     system('java Definition')
     chdir('..')
@@ -32,3 +42,4 @@ if __name__ == '__main__':
         download_antlr_jar()
     run_antlr()
     compile_java()
+    run_java()
